+++ src/types/employee/index.ts
@@ -1,5 +1,6 @@
 export * from './employee-status';
 export * from './employee';
+export * from './employee-lookup';
 export * from './departments';
 export * from './org-role';
 export * from './employee-create';